[New] Added test cases for record form form subforms subform header date
