+++ app/javascript/components/record-form/form/subforms/subform-header-date/component.jsx
@@ -32,7 +32,7 @@ const Component = ({ value, includeTime }) => {
     );
   }
 
-  return <span>{dateValue}</span>;
+  return <span data-testid="date-header">{dateValue}</span>;
 };
 
 Component.displayName = SUBFORM_HEADER_DATE;
